+++ src/full/Agda/TypeChecking/Rules/Decl.hs
@@ -160,8 +160,8 @@ checkDecl d = setCurrentRange d $ do
       A.ScopedDecl scope ds    -> none $ setScope scope >> mapM_ checkDeclCached ds
       A.FunDef i x delayed cs  -> impossible $ check x i $ checkFunDef delayed i x cs
       A.DataDef i x uc ps cs   -> impossible $ check x i $ checkDataDef i x uc ps cs
-      A.RecDef i x uc ind eta c ps tel cs -> impossible $ check x i $ do
-                                    checkRecDef i x uc ind eta c ps tel cs
+      A.RecDef i x uc ind eta pat c ps tel cs -> impossible $ check x i $ do
+                                    checkRecDef i x uc ind eta pat c ps tel cs
                                     blockId <- mutualBlockOf x
 
                                     -- Andreas, 2016-10-01 testing whether
@@ -408,8 +408,8 @@ highlight_ hlmod d = do
       highlight (A.Section i x tel [])
       when (hlmod == DoHighlightModuleContents) $ mapM_ (highlight_ hlmod) (deepUnscopeDecls ds)
     A.RecSig{}               -> highlight d
-    A.RecDef i x uc ind eta c ps tel cs ->
-      highlight (A.RecDef i x uc ind eta c A.noDataDefParams dummy cs)
+    A.RecDef i x uc ind eta pat c ps tel cs ->
+      highlight (A.RecDef i x uc ind eta pat c A.noDataDefParams dummy cs)
       -- The telescope has already been highlighted.
       where
       -- Andreas, 2016-01-22, issue 1790
@@ -449,7 +449,7 @@ checkPositivity_ mi names = Bench.billTo [Bench.Positivity] $ do
 --   for the old coinduction.)
 checkCoinductiveRecords :: [A.Declaration] -> TCM ()
 checkCoinductiveRecords ds = forM_ ds $ \case
-  A.RecDef _ q _ (Just (Ranged r CoInductive)) _ _ _ _ _ -> setCurrentRange r $ do
+  A.RecDef _ q _ (Just (Ranged r CoInductive)) _ _ _ _ _ _ -> setCurrentRange r $ do
     unlessM (isRecursiveRecord q) $ typeError $ GenericError $
       "Only recursive records can be coinductive"
   _ -> return ()
@@ -744,8 +744,9 @@ checkPragma r p =
           caseMaybeM (isRecord r) noRecord $ \case
             Record{ recInduction = ind, recEtaEquality' = eta } -> do
               unless (ind == Just CoInductive) $ noRecord
-              when (eta == Specified NoEta) $ typeError $ GenericError $
-                "ETA pragma conflicts with no-eta-equality declaration"
+              if | Specified NoEta{} <- eta -> typeError $ GenericError $
+                     "ETA pragma conflicts with no-eta-equality declaration"
+                 | otherwise -> return ()
             _ -> __IMPOSSIBLE__
           modifySignature $ updateDefinition r $ updateTheDef $ \case
             def@Record{} -> def { recEtaEquality' = Specified YesEta }